Sync upstream main through 393430e

+++ crates/codegen/xai-grok-tools/src/implementations/grok_build/bash/mod.rs
@@ -403,8 +403,8 @@ impl std::fmt::Display for KillReason {
 fn annotations(bash: &BashOutput) -> String {
     let mut s = String::new();
     if bash.truncated {
-        let shown = format_bytes(bash.output.len());
-        let total = format_bytes(bash.total_bytes);
+        let shown = format_bytes(bash.output.len() as u64);
+        let total = format_bytes(bash.total_bytes as u64);
         s.push_str(&format!(
             " [truncated: showing first/last {} of {} - full output at: {}]",
             shown, total, bash.output_file
@@ -436,8 +436,8 @@ pub(crate) fn format_default_prompt(bash: &BashOutput) -> String {
     let is_backgrounded = bash.signal.as_deref() == Some("backgrounded");
 
     if is_backgrounded {
-        let shown = format_bytes(bash.output.len());
-        let total = format_bytes(bash.total_bytes);
+        let shown = format_bytes(bash.output.len() as u64);
+        let total = format_bytes(bash.total_bytes as u64);
         format!(
             "[Command moved to background]\n\n\
              Partial output ({} of {} total):\n\n\
