+++ crates/codegen/xai-grok-tools/src/types/output.rs
@@ -1247,8 +1247,8 @@ impl xai_tool_runtime::ToolOutput for BashOutput {
         let mut stdout = String::from_utf8_lossy(&self.output).into_owned();
         let mut extra = serde_json::Map::new();
         if self.truncated {
-            let shown = crate::util::truncate::format_bytes(self.output.len());
-            let total = crate::util::truncate::format_bytes(self.total_bytes);
+            let shown = crate::util::truncate::format_bytes(self.output.len() as u64);
+            let total = crate::util::truncate::format_bytes(self.total_bytes as u64);
             stdout.push_str(&format!(
                 "\n[truncated: showing first/last {shown} of {total} - full output at: {}]",
                 self.output_file
